refactor(triggers): rename to bumpTriggerHealthCounter, skip retries on OOM
- Rename recordTriggerRun -> bumpTriggerHealthCounter (and its file
  trigger-run-recorder.ts -> trigger-health-counter.ts). The old name
  reads like it records one run; it actually bumps a per-piece daily
  aggregate counter. Reading it next to the new reportTriggerFailureAsFlowRun
  in the catch block, the two concerns are now visibly distinct: one
  ops counter, one customer-facing run row. The RPC method name on the
  wire contract is unchanged.

- The counter increment and the flow-run creation are independent RPCs.
  Run them concurrently in the catch block via Promise.all so trigger
  failures don't serialize two socket round-trips for no reason.

- Deterministic sandbox failures (OOM, sandbox timeout, log-size) will
  fail identically on retry — the input is the same and the sandbox
  behavior is deterministic. BullMQ's default 3-attempt retry on
  EXECUTE_WEBHOOK would produce three identical flow_run entries for
  every OOM'd webhook. Skip retries on those statuses (return OK); only
  the fallback INTERNAL_ERROR (potentially transient) still retries.
  This subsumes the previous isSandboxTimeout-specific skip.
