+++ packages/server/worker/src/lib/execute/utils/trigger-health-counter.ts
@@ -2,7 +2,7 @@ import { tryCatch } from '@activepieces/core-utils'
 import { ApLogger } from '@activepieces/server-utils'
 import { EngineResponseStatus, FlowTriggerType, FlowVersion, TriggerRunStatus, WorkerToApiContract } from '@activepieces/shared'
 
-export async function recordTriggerRun({ apiClient, log, flowVersion, platformId, status }: RecordTriggerRunParams): Promise<void> {
+export async function bumpTriggerHealthCounter({ apiClient, log, flowVersion, platformId, status }: BumpTriggerHealthCounterParams): Promise<void> {
     if (flowVersion.trigger.type !== FlowTriggerType.PIECE) {
         return
     }
@@ -10,11 +10,11 @@ export async function recordTriggerRun({ apiClient, log, flowVersion, platformId
     const triggerRunStatus = status === EngineResponseStatus.OK ? TriggerRunStatus.COMPLETED : TriggerRunStatus.FAILED
     const { error } = await tryCatch(() => apiClient.recordTriggerRun({ platformId, pieceName, status: triggerRunStatus }))
     if (error) {
-        log.warn({ error: String(error), piece: { name: pieceName }, flowVersion: { id: flowVersion.id } }, 'Failed to record trigger run stats')
+        log.warn({ error: String(error), piece: { name: pieceName }, flowVersion: { id: flowVersion.id } }, 'Failed to bump trigger health counter')
     }
 }
 
-type RecordTriggerRunParams = {
+type BumpTriggerHealthCounterParams = {
     apiClient: WorkerToApiContract
     log: ApLogger
     flowVersion: FlowVersion
